simplified toolbar by putting arbiter/player mode transition and recording mode toggle in new files
